refactor(lib,examples,devtools,claude): one type per quantity — a u8 sprite id of 256 silently overwrote sprite 0
API audit 2026-09-20, §3.4. Every change keeps the stack slot: no ABI change,
and callers passing narrower variables still convert.

Sprite ids. oamSet took a u16 id; oamSetX, oamSetY, oamSetXY, oamSetTile,
oamHide, oamDrawMeta and oamDrawMetaFlip took a u8. That was not only
inconsistent: a u8 parameter truncates 256 to 0 BEFORE the body's
`id >= MAX_SPRITES` check can refuse it, so oamSetX(256, ..) overwrote
sprite 0. All ids (and oamSetY / oamSetXY's y, and the metasprite return
value) are u16 now; the check sees the real value. Pinned by r_oam_id256.

fixLerp. `t` was a u8 documented as "0-256 = 0.0-1.0": 1.0 was unreachable,
256 was truncated to 0 and returned `a`. `t` is a u16; t >= 256 returns `b`
exactly (the 16x8 hardware multiply path cannot express it). Pinned by
r_lerp_t256 / r_lerp_t300.

dsp1Multiply is `s16 (s16, s16)` like every other signed DSP-1 quantity and
like DSP1_T_FROM_FIX, which produces its argument: as u16, `product < 0` was
always false. Pinned by r_mul_sign.

Negative control: with the previous lib the three main-fixture vectors fail
(0x000A / 0xA40E for the lerps, 0x9976 for the two overwritten sprites).

Stale doctrine retired. oamSet's header warned that coordinates MUST live in
an s16 struct because separate u16 variables "cause jerky movement" (a
compiler defect of early 2026), and continuous_scroll's comment repeated it as
REQUIRED. movement_move_sprite.toml pins plain-u16 motion to the pixel. The
port-example skill still told agents to bypass oamSet ("catastrophically
slow") by writing oamMemory[] by hand — which skips the Y-1 adjustment, the X
high bit and oam_max_id; the framesize cliff has been resolved since
2026-03.

Open in §3.4: typed map / object tables (a design, not a retype).

Validation: clean rebuild; diff_corpus 85/85 MATCH at +0; audio oracle 4/4
unchanged. WRAM oracle re-captured for 62 examples: lib code moved (wider
ids, the fixLerp guard), so the return addresses the oracle hashes off the
stack moved with it — rendering is identical everywhere. 208 / 30 / 16 / 10
fixture vectors; make tests, make lint, make docs-strict green.

+++ lib/include/snes/math.h
@@ -291,7 +291,10 @@ fixed fixClamp(fixed x, fixed min, fixed max);
  *
  * @param a Start value
  * @param b End value
- * @param t Interpolation factor (0-256 = 0.0-1.0)
+ * @param t Interpolation factor, 0-256 = 0.0-1.0. 256 and above return `b`
+ *          exactly. (`t` was a `u8` until 2026-09-21, which made the
+ *          documented 1.0 unreachable: 256 was truncated to 0 and returned `a`.
+ *          Same stack slot — no ABI change.)
  * @return Interpolated value
  *
  * @code
@@ -300,7 +303,7 @@ fixed fixClamp(fixed x, fixed min, fixed max);
  * fixed mid = fixLerp(start, end, 128);  // 50.0
  * @endcode
  */
-fixed fixLerp(fixed a, fixed b, u8 t);
+fixed fixLerp(fixed a, fixed b, u16 t);
 
 /*============================================================================
  * Square Root and Inverse Trigonometry
